remove unused file, split classes into their own files
