+++ Always On Server/Framework/ModData.cs
@@ -1,0 +1,12 @@
+﻿namespace Always_On_Server.Framework
+{
+    class ModData
+    {
+        public int FarmingLevel { get; set; }
+        public int MiningLevel { get; set; }
+        public int ForagingLevel { get; set; }
+        public int FishingLevel { get; set; }
+        public int CombatLevel { get; set; }
+        
+    }
+}
